requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
Babel==2.10.3
# Editable install with no version control (BittyTax==0.5.1.0a0)
-e c:\users\user\downloads\bittytax-master\bittytax-master
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
Django==4.1
django-adminlte3==0.1.6
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-multiselectfield==0.1.12
django-notifications-hq==1.7.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djongo==1.3.6
future==0.18.2
html5lib==1.1
idna==3.3
Jinja2==3.1.2
jsonfield==3.1.0
MarkupSafe==2.1.1
phonenumbers==8.12.52
Pillow==9.2.0
pymongo==3.12.3
PyPDF2==2.6.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
reportlab==3.6.6
requests==2.28.1
six==1.16.0
sqlparse==0.2.4
swapper==1.3.0
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.10
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==3.0.3
